requirements.txt
assemblyai
deep-translator
librosa
soundfile
pyrubberband
pydub
yandex-speechkit
pysndfx
moviepy
yandex-translate
TTS
audiostretchy